+++ CHANGELOG.md
@@ -12,6 +12,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - `hikari_pooling` config option to use HikariCP for database connections which are more stable
 - `discordsrv` integration added with a condition and an objective for account linking
 - `trade` objective to track trades
+- offset parameter for menu slots to support pagination
 ### Changed
 - `lore` of `item` placeholder can now be used without number to get all lines
 - `chat` notify io now only sends messages during a conversation if `bypassInterceptor` is set to `true`